+++ streams-tests/jvm/src/test/scala/zio/stream/InflateSpec.scala
