+++ streams/js/src/main/scala/zio/stream/platform.scala
@@ -1,4 +1,4 @@
-package zio.stream.experimental
+package zio.stream
 
 import zio._
 import zio.stacktracer.TracingImplicits.disableAutoTrace
