+++ streams/shared/src/main/scala-2.11-2.12/zio/stream/ZPipelineCompanionVersionSpecific.scala
