+++ streams/shared/src/main/scala-2.13+/zio/stream/ZPipelineCompanionVersionSpecific.scala
