+++ streams/shared/src/main/scala/zio/stream/Take.scala
