+++ streams/shared/src/main/scala/zio/stream/ZSink.scala
