+++ streams/shared/src/main/scala/zio/stream/internal/Utils.scala
